+++ Sources/TuistLoader/Models+ManifestMappers/Dependencies+ManifestMapper.swift
@@ -7,16 +7,13 @@ import TuistSupport
 
 extension TuistGraph.Dependencies {
     static func from(manifest: ProjectDescription.Dependencies) throws -> Self {
-        let carthageDependencies = try manifest.dependencies.reduce(into: [CarthageDependency]()) { result, dependency in
-            switch dependency {
-            case let .carthage(origin, requirement, platforms):
-                let origin = try TuistGraph.CarthageDependency.Origin.from(manifest: origin)
-                let requirement = try TuistGraph.CarthageDependency.Requirement.from(manifest: requirement)
-                let platforms = try platforms.map { try TuistGraph.Platform.from(manifest: $0) }
-                result.append(CarthageDependency(origin: origin, requirement: requirement, platforms: Set(platforms)))
+        let carthage: TuistGraph.CarthageDependencies? = try {
+            guard let carthage = manifest.carthage else {
+                return nil
             }
-        }
+            return try TuistGraph.CarthageDependencies.from(manifest: carthage)
+        }()
 
-        return Self(carthageDependencies: carthageDependencies)
+        return Self(carthage: carthage)
     }
 }