+++ fixtures/app_with_framework_and_tests_and_dependencies/Tuist/Dependencies.swift
@@ -1,5 +1,11 @@
 import ProjectDescription
 
-let dependencies = Dependencies([
-    .carthage(origin: .github(path: "Alamofire/Alamofire"), requirement: .exact("5.0.4"), platforms: [.macOS])
-])
+let dependencies = Dependencies(
+    carthage: .carthage(
+        [
+            .github(path: "Alamofire/Alamofire", requirement: .exact("5.0.4"))
+        ],
+        platforms: [.iOS, .macOS],
+        useXCFrameworks: true
+    )
+)